requirements.txt
aiofile==3.9.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.35.49
botocore==1.35.49
caio==0.9.17
cffi==1.17.1
channels==3.0.5
constantly==23.10.4
cryptography==43.0.3
daphne==3.0.2
deepgram-sdk==2.12.0
Django==5.1.2
djangorestframework==3.15.2
environs==11.0.0
frozenlist==1.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
marshmallow==3.23.0
multidict==6.1.0
packaging==24.1
propcache==0.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
yarl==1.16.0
zope.interface==7.1.1
